refactor(client-runtime): share voice supervisor tools
